ecommerce-solutions/requirements.txt
amqp
asgiref
billiard
boto3
botocore
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
coreapi
coreschema
cryptography
defusedxml
Django
djangorestframework-simplejwt
django-cors-headers
django-rest-auth
requests_oauthlib
django-storages
djangorestframework
djangorestframework-simplejwt
drf-yasg
dynaconf
fastadmin
gunicorn
h11
h2
hpack
hvac
Hypercorn
hyperframe
idna
inflection
isort
itypes
Jinja2
jmespath
kombu
loguru
MarkupSafe
packaging
pillow
priority
prompt_toolkit
pycparser
pydantic
PyJWT
python-dateutil
python-dotenv
python-slugify
python3-openid
psycopg2-binary
pytz
PyYAML
redis
requests
s3transfer
setuptools
shortuuid
six
sqlparse
text-unidecode
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
validators
vine
wcwidth
wsproto
djoser
celery